fix: ensure vcpkg installs packages explicitly before msbuild; fix warnings
- Add explicit 'vcpkg install --triplet x64-windows' step before msbuild
  so all packages (including opencv4) are reliably installed and any
  failure is visible with clear error output rather than cryptic C1083
- Fix C4018 in device_enum_dvp2.cpp: add static_cast<dvpUint32> for
  signed/unsigned comparison (max_camera_count int vs dvpUint32)
  and change loop variable i from int to dvpUint32
- Fix C4267 in common.cpp: add static_cast<int>(str.length()) to
  silence size_t -> int narrowing warning in is_utf8 call

+++ src/common/common.cpp
@@ -125,7 +125,7 @@ bool remove_files(const QString& dir, const QString& suffix)
 
 std::string utf8_to_ansi(const std::string& str)
 {
-    if (!is_utf8(str.c_str(), str.length()))
+    if (!is_utf8(str.c_str(), static_cast<int>(str.length())))
     {
         return str;
     }
